FitnessWebPython/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1

langchain
langchain-groq
langgraph
langchain-community

python-multipart==0.0.9
httpx==0.27.0